omit links that bypass explanation
I know this is maybe a painful cut, but it solves the biggest concern I have with this main landing page: someone clicks on jump_rr tools and never comes back to see the useful information (and tutorial) about those tools. If we instead have them go the Interactive tools page there is a probability that they see more context about the tools.